update emot dataloader and test
