+++ tests/test_nusantara.py
@@ -40,6 +40,7 @@ _TASK_TO_SCHEMA = {
     
     Tasks.SENTIMENT_ANALYSIS: "TEXT",
     Tasks.ASPECT_BASED_SENTIMENT_ANALYSIS: "TEXT",
+    Tasks.EMOTION_CLASSIFICATION: "TEXT",
 
     Tasks.DIALOGUE_SYSTEM: "KB"
 }
